Fix multiple HURUmap widgets on a single page
HURUmap widgets couldn't be shown on the same page because the script
wanted to process all of them at the same time.This only works on those
cases where some are not hidden as in hidden tabs.

This is fixed by making the script capable of process a single iframe at
a time as well as adding listeners for when inactive tabs on widgets and
indicators are activated.